split out Initialize method to extensions
provided Xamarin.Forms and Xamarin.iOS implementations
