+++ src/Sextant.XamForms.Tests/Sextant.XamForms.Tests.csproj
@@ -9,5 +9,6 @@
   <ItemGroup>
     <ProjectReference Include="..\Sextant.Mocks\Sextant.Mocks.csproj" />
     <ProjectReference Include="..\Sextant.XamForms\Sextant.XamForms.csproj" />
+    <ProjectReference Include="..\Sextant\Sextant.csproj" />
   </ItemGroup>
 </Project>
